requirements.txt
# Python project requirements
requests
psycopg2-binary
django
django-cors-headers
djangorestframework
django_redis
